refactor to keep better overview
